requirements.txt
numpy
opencv-python-headless
pygame-ce
